add t530

+++ release.sh
@@ -12,13 +12,14 @@ have_image_3=0
 have_image_4=0
 have_image_5=0
 have_image_6=0
+have_image_7=0
 
 usage()
 {
-        echo "Usage: $0 -v version -i img -f img -g img -h img -j img -k img"
+        echo "Usage: $0 -v version -i img -f img -g img -h img -j img -k img -l img"
 }
 
-args=$(getopt -o v:i:f:g:h:j:k: -- "$@")
+args=$(getopt -o v:i:f:g:h:j:k:l: -- "$@")
 if [ $? -ne 0 ] ; then
         usage
         exit 1
@@ -55,6 +56,11 @@ do
 	-k)
 		RELEASE_IMAGE_6=$2
 		have_image_6=1
+		shift
+		;;
+	-l)
+		RELEASE_IMAGE_7=$2
+		have_image_7=1
 		shift
 		;;
         -v)
@@ -102,6 +108,11 @@ if [ ! "$have_image_5" -gt 0 ] ; then
 	exit 1
 fi
 if [ ! "$have_image_6" -gt 0 ] ; then
+	echo "image missing"
+	usage
+	exit 1
+fi
+if [ ! "$have_image_7" -gt 0 ] ; then
 	echo "image missing"
 	usage
 	exit 1
@@ -184,6 +195,13 @@ if [ ! "$filesize" -eq "$reference_filesize" ] ; then
 	echo "filesize of release image is wrong"
 	exit 1
 fi
+filesize=$(wc -c <"${RELEASE_IMAGE_7}")
+reference_filesize=4194304
+if [ ! "$filesize" -eq "$reference_filesize" ] ; then
+	echo "filesize of release image is wrong"
+	exit 1
+fi
+
 
 RELEASE_DIR="skulls-${version}"
 rm -rf "$RELEASE_DIR"
@@ -196,6 +214,7 @@ cp "${RELEASE_IMAGE_3}" "$RELEASE_DIR"
 cp "${RELEASE_IMAGE_4}" "$RELEASE_DIR"
 cp "${RELEASE_IMAGE_5}" "$RELEASE_DIR"
 cp "${RELEASE_IMAGE_6}" "$RELEASE_DIR"
+cp "${RELEASE_IMAGE_7}" "$RELEASE_DIR"
 
 RELEASE_IMAGE_FILE=$(basename "${RELEASE_IMAGE}")
 cd ${RELEASE_DIR}
@@ -210,6 +229,8 @@ RELEASE_IMAGE_FILE_5=$(basename "${RELEASE_IMAGE_5}")
 sha256sum ${RELEASE_IMAGE_FILE_5} > "${RELEASE_IMAGE_FILE_5}.sha256"
 RELEASE_IMAGE_FILE_6=$(basename "${RELEASE_IMAGE_6}")
 sha256sum ${RELEASE_IMAGE_FILE_6} > "${RELEASE_IMAGE_FILE_6}.sha256"
+RELEASE_IMAGE_FILE_7=$(basename "${RELEASE_IMAGE_7}")
+sha256sum ${RELEASE_IMAGE_FILE_7} > "${RELEASE_IMAGE_FILE_7}.sha256"
 cd -
 
 # copy in device independent stuff
@@ -230,6 +251,14 @@ cp x230t/README.md \
 # copy in t430 stuff
 cp t430/README.md \
 	"$RELEASE_DIR/README_T430.md"
+
+# copy in t440p stuff
+cp t440p/README.md \
+	"$RELEASE_DIR/README_T440P.md"
+
+# copy in t530 stuff
+cp t530/README.md \
+	"$RELEASE_DIR/README_T530.md"
 
 tar -cJf "$RELEASE_DIR".tar.xz "$RELEASE_DIR"
